+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [[v0.4.0]]
 
+- add amm support
+  - Transaction models
+  - Transaction signing
+  - Request models
 - add sidechain support
   - Transaction models
   - Transaction signing